Bring theme to root component
