+++ screens/TreeMapScreen.js
@@ -1,0 +1,10 @@
+import * as React from "react";
+import { Text, StyleSheet } from "react-native";
+
+const TreeMapScreen = () => {
+  return <Text>Tree Map Screen</Text>;
+};
+
+const stylesheet = StyleSheet.create({});
+
+export default TreeMapScreen;
